+++ index.js
@@ -3,7 +3,7 @@
  * Apache-2.0
 */
 
-/* jshint node: true */
+/* eslint-env node */
 'use strict';
 
 var path = require('path');
@@ -17,11 +17,11 @@ module.exports = {
     return true;
   },
 
-  included(app) {
+  included( /* app */ ) {
     this._super.included.apply(this, arguments);
     // bundle scripts from vendor folder
     this.import('vendor/@esri/arcgis-rest-request/arcgis-rest-request.umd.js');
-    this.import('vendor/@esri/arcgis-rest-auth/arcgis-rest-auth.umd.js');
+    this.import('vendor/@esri/arcgis-rest-auth/auth.umd.js');
   },
 
   treeForVendor(vendorTree) {
@@ -30,8 +30,8 @@ module.exports = {
       destDir: '@esri/arcgis-rest-request'
     });
 
-    var arcgisAuthTree = new Funnel(path.dirname(require.resolve('@esri/arcgis-rest-auth/dist/umd/arcgis-rest-auth.umd.js')), {
-      files: ['arcgis-rest-auth.umd.js', 'arcgis-rest-auth.umd.js.map'],
+    var arcgisAuthTree = new Funnel(path.dirname(require.resolve('@esri/arcgis-rest-auth/dist/umd/auth.umd.js')), {
+      files: ['auth.umd.js', 'auth.umd.js.map'],
       destDir: '@esri/arcgis-rest-auth'
     });